Implement Column Sort  Fixes #1
Allow the user to sort by column by clicking on the column header. A
second click should reverse the order of sorting. Note: use the
database's ability to sort queries even though performance would
probably be better if the sorting was done locally. The point here is to
demonstrate the use of parameters in the API.